requirements.txt
amqp==2.3.2
asgiref==3.2.3
asn1crypto==0.24.0
bcrypt==3.1.7
beautifulsoup4==4.9.1
billiard==3.5.0.4
celery==4.2.1
certifi==2019.9.11
cffi==1.11.5
chardet==3.0.4
configparser==5.0.0
crontab==0.22.6
cryptography==2.7
cx-Oracle==8.0.0
cycler==0.10.0
Cython==0.29.14
Django==3.0
elasticsearch==1.9.0
happybase==1.2.0
html5lib==1.1
idna==2.7
kafka==1.3.5
kafka-python==1.4.7
kazoo==2.5.0
kiwisolver==1.0.1
kombu==4.2.1
matplotlib==2.2.3
minio==6.0.0
mysql-replication==0.21
ntlm-auth==1.5.0
numpy==1.15.1
paramiko==2.6.0
Pillow==6.2.0
ply==3.11
psutil==5.7.0
pycparser==2.18
pyelasticsearch==1.4.1
pykafka==2.8.0
pymongo==3.9.0
pymssql==2.1.4
PyMySQL==0.9.2
PyNaCl==1.3.0
pyparsing==2.2.0
python-dateutil==2.7.3
python3-pika==0.9.14
pytz==2018.5
pywinrm==0.4.1
redis==2.10.6
reportlab==3.5.32
requests==2.23.0
requests-ntlm==1.1.0
simplejson==3.16.0
six==1.11.0
soupsieve==2.0.1
sqlparse==0.3.0
tabulate==0.8.6
thriftpy==0.3.9
thriftpy2==0.4.10
tornado==3.1
tornado-celery==0.3.5
tornado-redis==2.4.18
urllib3==1.25.6
vine==1.1.4
webencodings==0.5.1
xlrd==1.2.0
xlwt==1.3.0
xmltodict==0.12.0
